Deler opp KompetanseUtil-testklassen til ulike temaer
