+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/eøs/kompetanse/util/TrekkFraKompetanser.kt
@@ -15,13 +15,13 @@ fun Kompetanse.trekkFra(oppdatertKompetanse: Kompetanse): Collection<Kompetanse>
             fom = gammelKompetanse.fom,
             tom = oppdatertKompetanse.fom?.minusMonths(1),
             barnAktørIder = oppdatertKompetanse.barnAktørIder
-        ).takeIf { it.fom != null && it.fom < it.tom }
+        ).takeIf { it.fom != null && it.fom <= it.tom }
 
     val kompetanseForEtterfølgendePerioder = gammelKompetanse.copy(
         fom = oppdatertKompetanse.tom?.plusMonths(1),
         tom = gammelKompetanse.tom,
         barnAktørIder = oppdatertKompetanse.barnAktørIder
-    ).takeIf { it.fom != null && it.fom < it.tom }
+    ).takeIf { it.fom != null && it.fom <= it.tom }
 
     return listOf(kompetanseForRestBarn, kompetanseForForegåendePerioder, kompetanseForEtterfølgendePerioder)
         .filterNotNull()
@@ -30,4 +30,12 @@ fun Kompetanse.trekkFra(oppdatertKompetanse: Kompetanse): Collection<Kompetanse>
 fun Iterable<Kompetanse>.trekkFra(skalFjernes: Iterable<Kompetanse>) =
     this.flatMap { kompetanse ->
         skalFjernes.find { kompetanse.inneholder(it) }?.let { kompetanse.trekkFra(it) } ?: listOf(kompetanse)
+    }
+
+fun Iterable<Kompetanse>.trekkFra(skalFjernes: Kompetanse) =
+    this.flatMap { kompetanse ->
+        if (kompetanse.inneholder(skalFjernes))
+            kompetanse.trekkFra(skalFjernes)
+        else
+            listOf(kompetanse)
     }
